+++ amqp-kit/client.go
@@ -13,7 +13,8 @@ import (
 	"github.com/streadway/amqp"
 )
 
-const defaultReconnectAfterDuration = 2 * time.Second
+const defaultReconnectAfterDuration = 500 * time.Millisecond
+const defaultWaitWorkerDuration = 5 * time.Second
 
 // Publisher interface use for publish amqp - message
 type Publisher interface {
@@ -52,6 +53,7 @@ type Config struct {
 	ChannelPoolSize        int
 	ChannelRetryCount      int
 	ReconnectAfterDuration time.Duration
+	WaitWorkerDuration     time.Duration
 }
 
 // New AMQP Client with connection
@@ -61,7 +63,11 @@ func New(cfg *Config) (*Client, error) {
 		stopClientChan: make(chan struct{}),
 	}
 
-	return ser, ser.reconnect()
+	if err := ser.reconnect(); err != nil {
+		return nil, err
+	}
+
+	return ser, nil
 }
 
 // MakeDsn - creates dsn from config
@@ -139,7 +145,12 @@ func (c *Client) Serve(si []SubscribeInfo) (err error) {
 		subscribers[si.Queue] = &s
 	}
 
-	for _, sub := range subscribers {
+	workerDuration := c.config.WaitWorkerDuration
+	if workerDuration == 0 {
+		workerDuration = defaultWaitWorkerDuration
+	}
+
+	for q, sub := range subscribers {
 		for i := 0; i < sub.Workers; i++ {
 			go func(si *SubscribeInfo) {
 				for {
@@ -157,9 +168,33 @@ func (c *Client) Serve(si []SubscribeInfo) (err error) {
 				}
 			}(sub)
 		}
+
+		t := time.Now().Add(workerDuration)
+		for {
+			conn := c.getConnection()
+			ch, err := conn.getChan()
+			if err != nil {
+				return fmt.Errorf("AMQP: Channel err: %s", err.Error())
+			}
+
+			if time.Now().After(t) {
+				return fmt.Errorf(`worker wait timeout`)
+			}
+
+			// the channel will be closed if error
+			qu, err := ch.c.QueueInspect(q)
+			if err != nil {
+				continue
+			}
+			conn.putChan(ch)
+
+			if qu.Consumers == sub.Workers {
+				break
+			}
+		}
 	}
 
-	return
+	return nil
 }
 
 func (c *Client) receive(si *SubscribeInfo) error {
@@ -176,7 +211,7 @@ func (c *Client) receive(si *SubscribeInfo) error {
 
 	msgs, err := ch.c.Consume(si.Queue, si.Name, false, false, false, false, nil)
 	if err != nil {
-		return fmt.Errorf(" Channel consume err: %s", err.Error())
+		return fmt.Errorf("Channel consume err: %s ", err.Error())
 	}
 	fun := NewSubscriber(si.E, si.Dec, si.Enc, si.O...).ServeDelivery(ch.c)
 
@@ -190,7 +225,7 @@ func (c *Client) receive(si *SubscribeInfo) error {
 		fun(&d)
 	}
 
-	return fmt.Errorf(" Close channel error ")
+	return fmt.Errorf("Close channel error ")
 }
 
 // DeclareAndBind create exchange, queue and create bind by key